:recycle: refactor code
