+++ setup.py
@@ -50,7 +50,7 @@ setup(
     ],
     entry_points={
         "console_scripts": [
-            "run_disdrodb_l0_reader=disdrodb.L0.run_disdrodb_l0_reader:run_reader",
+            "run_disdrodb_l0_reader=disdrodb.L0.run_disdrodb_l0_reader:run_reader_cmd",
         ]
     },
 )
